+++ src/hooks/preloadEmoji.ts
@@ -1,4 +1,4 @@
-import { GetEmojiUrl } from '../components/emoji/Emoji';
+import { GetEmojiUrl } from '../components/emoji/BaseEmojiProps';
 import { DataEmoji } from '../dataUtils/DataTypes';
 import { emojiUnified, emojiVariations } from '../dataUtils/emojiSelectors';
 import { EmojiStyle } from '../types/exposedTypes';
